Fixed first iteration of Linux warnings

+++ game/miniutl/fmtstr.h
@@ -215,12 +215,13 @@ public:
 	}
 	inline const char* SetInt64(int64 n64)
 	{
-		NUMSTR_CHECK_FAST(n64, uint64) m_nLength = V_snprintf(m_szBuf, sizeof(m_szBuf), "%lld", n64);
+		NUMSTR_CHECK_FAST(n64, uint64) m_nLength = V_snprintf(m_szBuf, sizeof(m_szBuf), "%lld", (long long int)n64);
 		return m_szBuf;
 	}
 	inline const char* SetUint64(uint64 un64)
 	{
-		NUMSTR_CHECK_FAST(un64, uint64) m_nLength = V_snprintf(m_szBuf, sizeof(m_szBuf), "%llu", un64);
+		NUMSTR_CHECK_FAST(un64, uint64)
+		m_nLength = V_snprintf(m_szBuf, sizeof(m_szBuf), "%llu", (long long unsigned int)un64);
 		return m_szBuf;
 	}
 #endif
